Update Value Classes With Validation and add casses for all the validations
